fix test invocation for linux #3
