+++ package.json
@@ -90,7 +90,7 @@
     "pretest": "npm run compile && npm run lint",
     "lint": "eslint src --ext ts",
     "test": "node ./scripts/runTests.js",
-    "test:coverage": "npm run compile && npm run lint && node ./scripts/runTestsWithCoverage.js",
+    "test:coverage": "npm run compile && npm run lint && node ./scripts/runTests.js",
     "package": "npx @vscode/vsce package"
   },
   "devDependencies": {
