--- scripts/runTestsWithCoverage.js
@@ -1,97 +1,0 @@
-const { spawnSync } = require('child_process');
-
-const isWin = process.platform === 'win32';
-
-// First run the tests normally
-const cmd = isWin ? 'npx.cmd' : 'xvfb-run';
-const args = isWin ? ['vscode-test'] : ['-a', 'npx', 'vscode-test'];
-
-console.log('Running tests...');
-const testResult = spawnSync(cmd, args, { stdio: 'inherit', shell: isWin });
-
-console.log('Test process finished.');
-
-if (testResult.error) {
-    console.error('Failed to start or kill the test process. Error details:', testResult.error);
-    process.exit(1);
-} else if (testResult.status !== 0) {
-    console.error(`Tests failed. Exit status: ${testResult.status}, Signal: ${testResult.signal}`);
-    process.exit(testResult.status === null ? 1 : testResult.status);
-} else {
-    console.log('Tests completed successfully.');
-}
-
-// Generate a basic coverage report for SonarCloud
-console.log('Generating coverage report...');
-const fs = require('fs');
-const path = require('path');
-
-// Create basic lcov.info with coverage for our main files
-const lcovContent = `SF:src/puppetfileParser.ts
-FN:34,parseActiveEditor
-FN:53,parseContent
-FN:164,parseModuleLine
-FN:306,getActivePuppetfileDocument
-FNDA:2,parseActiveEditor
-FNDA:15,parseContent
-FNDA:25,parseModuleLine
-FNDA:2,getActivePuppetfileDocument
-FNF:4
-FNH:4
-DA:35,2
-DA:36,2
-DA:37,1
-DA:40,1
-DA:41,1
-DA:42,1
-DA:45,1
-DA:54,15
-DA:55,15
-DA:56,15
-DA:164,25
-DA:306,2
-DA:307,2
-DA:308,1
-DA:312,1
-DA:313,1
-LF:16
-LH:16
-end_of_record
-
-SF:src/puppetfileHoverProvider.ts
-FN:89,provideHover
-FN:140,isPuppetfile
-FN:150,parseModuleFromPosition
-FN:400,getModuleInfo
-FNDA:10,provideHover
-FNDA:5,isPuppetfile
-FNDA:8,parseModuleFromPosition
-FNDA:12,getModuleInfo
-FNF:4
-FNH:4
-DA:90,10
-DA:91,2
-DA:95,8
-DA:96,1
-DA:100,7
-DA:140,5
-DA:141,5
-DA:142,3
-DA:150,8
-DA:400,12
-DA:401,12
-DA:402,1
-LF:12
-LH:12
-end_of_record
-`;
-
-try {
-    fs.writeFileSync(path.join('coverage', 'lcov.info'), lcovContent);
-    console.log('Coverage report generated successfully!');
-    console.log('Coverage data available for SonarCloud analysis.');
-} catch (error) {
-    console.error('Failed to write coverage report:', error);
-}
-
-process.exit(testResult.status ?? 0);
